更新

+++ options.go
@@ -6,9 +6,9 @@ type Options struct {
 	// 生成的mock数据的最小长度
 	MinLen int
 	// 生成的mock数据的最大值
-	Max int
+	Max int64
 	// 生成的mock数据的最小值
-	Min int
+	Min int64
 }
 
 type Option func(*Options)
@@ -25,13 +25,13 @@ func WithMinLen(minLen int) Option {
 	}
 }
 
-func WithMax(max int) Option {
+func WithMax(max int64) Option {
 	return func(o *Options) {
 		o.Max = max
 	}
 }
 
-func WithMin(min int) Option {
+func WithMin(min int64) Option {
 	return func(o *Options) {
 		o.Min = min
 	}
